+++ src/Admin/AdminHealth.php
@@ -53,26 +53,26 @@ class AdminHealth {
 		$fields = array();
 
 		// License key.
-		$fields['pronamic_pay_license_key'] = array(
+		$fields['license_key'] = array(
 			'label'   => __( 'Support license key', 'pronamic_ideal' ),
 			'value'   => esc_html( get_option( 'pronamic_pay_license_key', __( 'No license key found', 'pronamic_ideal' ) ) ),
 			'private' => true,
 		);
 
 		// License status.
-		$fields['pronamic_pay_license_status'] = array(
+		$fields['license_status'] = array(
 			'label' => __( 'License status', 'pronamic_ideal' ),
 			'value' => esc_html( $this->plugin->license_manager->get_formatted_license_status() ),
 		);
 
 		// Next scheduled license check.
-		$fields['pronamic_pay_next_license_check'] = array(
+		$fields['next_license_check'] = array(
 			'label' => __( 'Next scheduled license check', 'pronamic_ideal' ),
 			'value' => esc_html( $this->plugin->license_manager->get_formatted_next_license_check() ),
 		);
 
 		// Time.
-		$fields['pronamic_pay_time'] = array(
+		$fields['time'] = array(
 			'label' => __( 'Time (UTC)', 'pronamic_ideal' ),
 			'value' => esc_html( gmdate( __( 'Y/m/d g:i:s A', 'pronamic_ideal' ) ) ),
 		);
@@ -84,9 +84,15 @@ class AdminHealth {
 			$openssl_version = OPENSSL_VERSION_TEXT;
 		}
 
-		$fields['pronamic_pay_openssl_version'] = array(
+		$fields['openssl_version'] = array(
 			'label' => __( 'OpenSSL version', 'pronamic_ideal' ),
 			'value' => esc_html( $openssl_version ),
+		);
+
+		// Active plugin integrations.
+		$fields['active_plugin_integrations'] = array(
+			'label' => __( 'Active plugin integrations', 'pronamic_ideal' ),
+			'value' => $this->get_active_plugin_integrations_debug(),
 		);
 
 		// Add debug information section.
@@ -96,6 +102,33 @@ class AdminHealth {
 		);
 
 		return $debug_information;
+	}
+
+	/**
+	 * Get active plugin integrations debug.
+	 *
+	 * @return string
+	 */
+	private function get_active_plugin_integrations_debug() {
+		$active = array();
+
+		// Check integrations.
+		foreach ( $this->plugin->plugin_integrations as $integration ) {
+			if ( ! $integration->is_active() ) {
+				continue;
+			}
+
+			$active[] = $integration->get_name();
+		}
+
+		// Default result no active integrations.
+		if ( empty( $active ) ) {
+			$active[] = __( 'None', 'pronamic_ideal' );
+		}
+
+		$result = \implode( ', ', $active );
+
+		return $result;
 	}
 
 	/**